added universal components RecipeCard and RecipeList, started doing myRecipePage
