+++ src/components/universalComponents/RecipesList/RecipesList.styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const RecipeList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  gap: 18px;
+
+  @media screen and (min-width: 768px) {
+    gap: 40px;
+  }
+
+  @media screen and (min-width: 768px) {
+    gap: 50px;
+  }
+`;
